update: 명단관리 > 요일 순서 정렬 및 리팩토링
