+++ client/src/app/list-management/_components/AttendanceItem.tsx
@@ -44,12 +44,29 @@ const AttendanceItem = (props: PropsType) => {
         SATURDAY: '토',
         SUNDAY: '일',
     };
+    // 요일 순서
+    const dayOrder = [
+        'MONDAY',
+        'TUESDAY',
+        'WEDNESDAY',
+        'THURSDAY',
+        'FRIDAY',
+        'SATURDAY',
+        'SUNDAY',
+    ];
+
+    // 요일 순서 정렬을 위한 비교 함수
+    const compareDays = (a: ScheduleType, b: ScheduleType) => {
+        return dayOrder.indexOf(a.day) - dayOrder.indexOf(b.day);
+    };
 
     const [attendanceDay, setAttendanceDay] = useState<string>('');
 
     useEffect(() => {
         if (!_.isEmpty(item.schedules)) {
             const attendanceDays: string[] = [];
+
+            item.schedules.sort(compareDays);
 
             item.schedules.forEach((schedule: ScheduleType) => {
                 const day = days[schedule.day];
